// web-ui/src/api/models/ApiV1ProfilesPostRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * JobCounselor API
 * Minimal API surface for the JobCounselor backend
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ApiV1ProfilesPostRequest
 */
export interface ApiV1ProfilesPostRequest {
    /**
     * 
     * @type {string}
     * @memberof ApiV1ProfilesPostRequest
     */
    fullName?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiV1ProfilesPostRequest
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiV1ProfilesPostRequest
     */
    phone?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiV1ProfilesPostRequest
     */
    summary?: string;
}

/**
 * Check if a given object implements the ApiV1ProfilesPostRequest interface.
 */
export function instanceOfApiV1ProfilesPostRequest(value: object): value is ApiV1ProfilesPostRequest {
    return true;
}

export function ApiV1ProfilesPostRequestFromJSON(json: any): ApiV1ProfilesPostRequest {
    return ApiV1ProfilesPostRequestFromJSONTyped(json, false);
}

export function ApiV1ProfilesPostRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ApiV1ProfilesPostRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'fullName': json['fullName'] == null ? undefined : json['fullName'],
        'email': json['email'] == null ? undefined : json['email'],
        'phone': json['phone'] == null ? undefined : json['phone'],
        'summary': json['summary'] == null ? undefined : json['summary'],
    };
}

export function ApiV1ProfilesPostRequestToJSON(json: any): ApiV1ProfilesPostRequest {
    return ApiV1ProfilesPostRequestToJSONTyped(json, false);
}

export function ApiV1ProfilesPostRequestToJSONTyped(value?: ApiV1ProfilesPostRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'fullName': value['fullName'],
        'email': value['email'],
        'phone': value['phone'],
        'summary': value['summary'],
    };
}